requirements.txt
Flask~=2.2.2
mysql-connector-python~=8.0.32
Faker~=16.6.0